Management commands for working with swarm nodes

Methods for managing docker swarm nodes. This object is $node within a docker_client object.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

node_delete

Delete a node. Similar to the cli command docker node rm. Usage:
  node_delete(id, force = NULL)
Arguments:

id: The ID or name of the node

force: Force remove a node from the swarm

get

Get a service by name or id Usage:
  get(id)
Arguments:

id: A scalar character with the network's name or id (abbreviations of the id are allowed and will be resolved by the docker daemon).

help

Display help for this object Usage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

node_list

List nodes. Similar to the cli command docker node list. Usage:
  node_list(filters = NULL)
Arguments:

filters: Filters to process on the nodes list, as a named character vector.

id=<node id>

label=<engine label>

membership=(accepted|pending)

name=<node name>

role=(manager|worker)

See also

docker_swarm_collection for management commands for the swarm itself.