Management commands for working with a particular docker plugin

Methods for working with a particular docker plugin. Plugin objects are returned by installing or building a docker plugin, or by using $plugin$get to fetch an existing plugin by name or id.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues

Methods

plugin_configure

Configure a plugin. Similar to the cli command docker plugin set.

Usage:

  plugin_configure(body = NULL)

Arguments:

body: NA

plugin_disable

Disable a plugin. Similar to the cli command docker plugin disable.

Usage:

  plugin_disable()

plugin_enable

Enable a plugin. Similar to the cli command docker plugin enable.

Usage:

  plugin_enable(timeout = 0L)

Arguments:

timeout: Set the HTTP client timeout (in seconds)

help

Display help for this object

Usage:

  help(help_type = getOption("help_type"))

Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

id

Return the plugins' id

Usage:

  id()

inspect

Return detailed information about this plugin

Usage:

  inspect(reload = TRUE)

Arguments:

reload: Logical scalar, indicating if the information should be refreshed from the daemon. This is useful to set to FALSE after a plugin has been removed.

is_enabled

Test if plugin is enabled

Usage:

  is_enabled(reload = TRUE)

Arguments:

reload: Logical scalar, indicating if the information should be refreshed from the daemon. This is useful to set to FALSE after a plugin has been removed.

name

Return the plugin's name

Usage:

  name()

reload

Refresh infomation on the plugin from the server, returning $inspect() invisibly.

Usage:

  reload()

plugin_remove

Remove a plugin. Similar to the cli command docker plugin rm.

Usage:

  plugin_remove(force = NULL)

Arguments:

force: Disable the plugin before removing. This may result in issues if the plugin is in use by a container.

See also