Management commands for working with docker plugins

Methods for working with docker plugins. This object is $plugin within a docker_client object.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues

Methods

plugin_create

Create a plugin. Similar to the cli command docker plugin create.

Usage:

  plugin_create(name, plugin_data_dir)

Arguments:

name: The name of the plugin. The :latest tag is optional, and is the default if omitted.

plugin_data_dir: Path to tar containing plugin rootfs and manifest

get

Get a plugin by name or id

Usage:

  get(name)

Arguments:

name: A scalar character with the plugins' name.

help

Display help for this object

Usage:

  help(help_type = getOption("help_type"))

Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

plugin_install

Install a plugin. Pulls and installs a plugin. After the plugin is installed, it can be enabled using the hrefhttps://docs.docker.com/engine/api/1.29/#operation/PostPluginsEnablePOST /plugins/{name}/enable endpoint. Similar to the cli command docker plugin install.

Usage:

  plugin_install(remote, alias = NULL, registry_auth = NULL,
      disable = FALSE, grant_all = NULL, stream = stdout())

Arguments:

remote: Remote reference for plugin to install.

The :latest tag is optional, and is used as the default if omitted.

alias: Local name for the pulled plugin.

The :latest tag is optional, and is used as the default if omitted.

registry_auth: A base64-encoded auth configuration to use when pulling a plugin from a registry. hrefhttps://docs.docker.com/engine/api/1.29/#section/AuthenticationSee the authentication section for details.

disable: Do not enable the plugin on install

grant_all: Logical, indicating if all requested permissions should be granted. If NULL, this will prompt interactively for permission. Setting this to FALSE will cause the installation to fail.

stream: The stream to send output to. Options here are (1) a connection object (e.g. stdout(), stderr() or a writable open file connection object, (2) a scalar character indicating a filename to write to, or (3) FALSE or NULL to disable any output.

plugin_list

List plugins. Returns information about installed plugins. Similar to the cli command docker plugin ls.

Usage:

  plugin_list(filters = NULL)

Arguments:

filters: A JSON encoded value of the filters (a map[string][]string) to process on the plugin list. Available filters:

capability=<capability name>

enable=<true>|<false>

plugin_privileges

Get plugin privileges

Usage:

  plugin_privileges(remote)

Arguments:

remote: The name of the plugin. The :latest tag is optional, and is the default if omitted.

See also