Management commands for working with swarm secrets

Methods for managing docker swarm secrets. This object is $secret within a docker_client object.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

secret_create

Create a secret. Similar to the cli command docker secret create.Usage:
  secret_create(name, data, labels = NULL)
Arguments:

name: User-defined name of the secret.

data: Secret data, as a raw vector or character string

labels: User-defined key/value metadata.

help

Display help for this objectUsage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

secret_inspect

Inspect a secret. Similar to the cli command docker secret inspect.Usage:
  secret_inspect(id)
Arguments:

id: ID of the secret

secret_list

List secrets. Similar to the cli command docker secret ls.Usage:
  secret_list(filters = NULL)
Arguments:

filters: A named character vector of filters to process on the secrets list. Available filters:

id=<secret id>

label=<key> or label=<key>=value

name=<secret name>

names=<secret name>

secret_delete

Delete a secret. Similar to the cli command docker secret rm.Usage:
  secret_delete(id)
Arguments:

id: ID of the secret

secret_update

Update a SecretUsage:
  secret_update(id, version, name = NULL, labels = NULL,
      data = NULL)
Arguments:

id: The ID or name of the secret

version: The version number of the secret object being updated. This is required to avoid conflicting writes.

name: User-defined name of the secret.

labels: User-defined key/value metadata.

data: Secret data, as a raw vector or character string

See also