Management commands for working with a particular docker volume

Methods for working with a particular docker volume. Volume objects are returned by creating a docker volume, or by using $volume$get to fetch an existing volume by name or id.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

help

Display help for this objectUsage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

inspect

Return detailed information about this volume. Similar to the cli command docker volume inspect.Usage:
  inspect(reload = TRUE)
Arguments:

reload: Logical scalar, indicating if the information should be refreshed from the daemon. This is useful to set to FALSE after an volume has been removed.

map

Generate a path suitable to be passed in to docker$container$run or docker$container$createUsage:
  map(path, readonly = FALSE)
Arguments:

path: Path in the container to map the volume to

readonly: Logical, indicating if this should be a readonly mount

name

Return name for this volumeUsage:
  name()

reload

Refresh infomation on the volume from the server, returning $inspect() invisibly.Usage:
  reload()

volume_delete

Remove a volume. Instruct the driver to remove the volume. Similar to the cli command docker volume rm.Usage:
  volume_delete(force = NULL)
Arguments:

force: Force the removal of the volume

See also