Test if docker available

Test if we can construct a docker client and ping it. This is intended to help in debug connection issues, and also for use in tests. For example, you might implement a testthat skip test that skips if stevedore::docker_available() returns FALSE to conditionally use stevedore/docker within tests.

docker_available(api_version = NULL, url = NULL, ...,
  http_client_type = NULL, verbose = FALSE)

Arguments

api_version
Version of the API request from the api. Options are NULL (the default) - use the package's default version (currently 1.29 ), a version as a string or numeric_version object (supported between 1.25 and 1.37 ), or the string detect which will use the highest version out of the version reported by the api and 1.37 .

url
The URL for the docker daemon. This can be an absolute file path (for a unix socket on macOS/Linux), a named pipe (e.g., npipe:////./pipe/docker_engine) on Windows, or (eventually) an http or https url (e.g., https://localhost:8888), though this is not yet supported.

...
Reserved for future use. Passing in any unrecognised argument will throw an error.

http_client_type
HTTP client type to use. The options are (currently) "curl", which uses the curl package (works over unix sockets and eventually over TCP) and httppipe which works over unix sockets and eventually windows named pipes, using the Docker SDK's pipe code via the httppipe package. Not all functionality is supported with the httppipe client. This option may eventually be moved into the ... argument as is not intended for end-user use; it is primarily intended for debugging in development (forcing the httppipe client where the curl client would ordinarily be preferred).

verbose
Logical, indicating if information should be printed about failures to connect. If FALSE (the default) the function runs silently.

Value

Logical scalar, TRUE if docker_client(...) would succeed.

Details

Reasons for failure to connect might include:

You do not have docker installed

You have docker installed but the socket in a nonstandard place and have not adjusted environment variables accordingly

You do not have permission to write to the docker socket

You are on windows and the required python packages to get everything working there are not present or configured correctly.

Examples

# Is docker available on your system?
stevedore::docker_available()

#> [1] TRUE