Create docker client

Create a docker client object

docker_client(api_version = NULL, url = NULL, ...,
  http_client_type = NULL, quiet = FALSE)

Arguments

api_version
Version of the API request from the api. Options are NULL (the default) - use the package's default version (currently 1.29 ), a version as a string or numeric_version object (supported between 1.25 and 1.37 ), or the string detect which will use the highest version out of the version reported by the api and 1.37 .

url
The URL for the docker daemon. This can be an absolute file path (for a unix socket on macOS/Linux), a named pipe (e.g., npipe:////./pipe/docker_engine) on Windows, or (eventually) an http or https url (e.g., https://localhost:8888), though this is not yet supported.

...
Reserved for future use. Passing in any unrecognised argument will throw an error.

http_client_type
HTTP client type to use. The options are (currently) "curl", which uses the curl package (works over unix sockets and eventually over TCP) and httppipe which works over unix sockets and eventually windows named pipes, using the Docker SDK's pipe code via the httppipe package. Not all functionality is supported with the httppipe client. This option may eventually be moved into the ... argument as is not intended for end-user use; it is primarily intended for debugging in development (forcing the httppipe client where the curl client would ordinarily be preferred).

quiet
Suppress informational messages.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Management commands

config

Manage config - see docker_config_collection

container

Manage container - see docker_container_collection

image

Manage image - see docker_image_collection

network

Manage network - see docker_network_collection

node

Manage node - see docker_node_collection

plugin

Manage plugin - see docker_plugin_collection

secret

Manage secret - see docker_secret_collection

service

Manage service - see docker_service_collection

swarm

Manage swarm - see docker_swarm_collection

task

Manage task - see docker_task_collection

types

Manage types - see docker_type_collection

volume

Manage volume - see docker_volume_collection

Methods

api_version

Return the version of the API that the docker client is using to communicate with the server, as a string (e.g., "1.29") Usage:
  api_version()

system_df

Get data usage information. Similar to the cli command docker system df. Usage:
  system_df()

system_events

Monitor events. Stream real-time events from the server. Various objects within Docker report events when something happens to them. Containers report these events: attach, commit, copy, create, destroy, detach, die, exec_create, exec_detach, exec_start, export, health_status, kill, oom, pause, rename, resize, restart, start, stop, top, unpause, update Images report these events: delete, import, load, pull, push, save, tag, untag Volumes report these events: create, mount, unmount, destroy Networks report these events: create, connect, disconnect, destroy The Docker daemon reports these events: reload. Similar to the cli command docker events. Usage:
  system_events(since = NULL, until = NULL, filters = NULL)
Arguments:

since: Show events created since this timestamp then stream new events.

until: Show events created until this timestamp then stop streaming.

filters: A named character vector of filters to process on the event list. Available filters:

container=<string> container name or ID

daemon=<string> daemon name or ID

event=<string> event type

image=<string> image name or ID

label=<string> image or container label

network=<string> network name or ID

plugin=<string> plugin name or ID

type=<string> object to filter by, one of container, image, volume, network, or daemon

volume=<string> volume name or ID

help

Display help for this object Usage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

system_info

Get system information. Similar to the cli command docker info. Usage:
  system_info()

system_auth

Check auth configuration. Validate credentials for a registry and, if available, get an identity token for accessing the registry without password. Similar to the cli command docker login. Usage:
  system_auth(username = NULL, password = NULL, email = NULL,
      serveraddress = NULL)
Arguments:

username: For authentication to check

password: For authentication to check

email: For authentication to check

serveraddress: For authentication to check

system_ping

Ping. This is a dummy endpoint you can use to test if the server is accessible. Usage:
  system_ping()

system_version

Get version. Returns the version of Docker that is running and various information about the system that Docker is running on. Similar to the cli command docker version. Usage:
  system_version()