Management commands for working with swarm configs

Methods for managing docker swarm configs. This object is $config within a docker_client object.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

config_create

Create a config. Unimplemented, as 'config_create' (POST /configs/create) requires docker API version at least 1.30 (version 1.29 used) Usage:
  config_create(...)
Arguments:

...: Ignored in this version

help

Display help for this object Usage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

config_inspect

Inspect a config. Unimplemented, as 'config_inspect' (GET /configs/id) requires docker API version at least 1.30 (version 1.29 used) Usage:
  config_inspect(...)
Arguments:

...: Ignored in this version

config_list

List configs. Unimplemented, as 'config_list' (GET /configs) requires docker API version at least 1.30 (version 1.29 used) Usage:
  config_list(...)
Arguments:

...: Ignored in this version

config_delete

Delete a config. Unimplemented, as 'config_delete' (DELETE /configs/id) requires docker API version at least 1.30 (version 1.29 used) Usage:
  config_delete(...)
Arguments:

...: Ignored in this version

config_update

Update a Config. Unimplemented, as 'config_update' (POST /configs/id/update) requires docker API version at least 1.30 (version 1.29 used) Usage:
  config_update(...)
Arguments:

...: Ignored in this version

See also

docker_swarm_collection for management commands for the swarm itself, and docker_secret_collection for a similar inteface for configuring sensitive configurations.