Commands for working with a docker exec instance

Methods for working with docker "exec" instances, which are returned by running exec on a running container

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

help

Display help for this objectUsage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

id

Return the exec instance's idUsage:
  id()

inspect

Return detailed information about this exec instanceUsage:
  inspect(reload = TRUE)
Arguments:

reload: Logical scalar, indicating if the information should be refreshed from the daemon.

reload

Refresh infomation on the exec instance from the server, returning $inspect() invisibly.Usage:
  reload()

exec_resize

Resize an exec instance. Resize the TTY session used by an exec instance. This endpoint only works if tty was specified as part of creating and starting the exec instance.Usage:
  exec_resize(h = NULL, w = NULL)
Arguments:

h: Height of the TTY session in characters

w: Width of the TTY session in characters

exec_start

Start an exec instance. Starts a previously set up exec instance. If detach is true, this endpoint returns immediately after starting the command. Otherwise, it sets up an interactive session with the command.Usage:
  exec_start(detach = FALSE, tty = NULL, stream = stdout())
Arguments:

detach: Detach from the command.

tty: Allocate a pseudo-TTY.

stream: The stream to send output to. Options here are (1) a connection object (e.g. stdout(), stderr() or a writable open file connection object, (2) a scalar character indicating a filename to write to, or (3) FALSE or NULL to disable any output.

See also