Management commands for working with docker volumes

Methods for working with docker volumes. This object is $volume within a docker_client object.

Details

Below is reference documentation for all methods for version '1.29' of the docker API - other versions are available. This documentation is automatically generated from docker's API schema, and so inaccuracies may exist between it and stevedore's interface (especially references to JSON objects). Please report any documentation that might be improved at https://github.com/richfitz/stevedore/issues
Methods

volume_create

Create a volume. Similar to the cli command docker volume create. Usage:
  volume_create(name = NULL, driver = NULL, driver_opts = NULL,
      labels = NULL)
Arguments:

name: The new volume's name. If not specified, Docker generates a name.

driver: Name of the volume driver to use.

driver_opts: A mapping of driver options and values. These options are passed directly to the driver and are driver specific.

labels: User-defined key/value metadata.

get

Get a volume by name or id Usage:
  get(name)
Arguments:

name: A scalar character with the volume's name.

help

Display help for this object Usage:
  help(help_type = getOption("help_type"))
Arguments:

help_type: Passed to utils::help, can be one of "text", "html" or "pdf" (or an abbreviation). By default it uses the value getOption("help_type") and should follow the same behaviour as other R help (e.g., using "?")

volume_list

List volumes. Similar to the cli command docker volume ls. Usage:
  volume_list(filters = NULL)
Arguments:

filters: A named character vector of filters to process on the volumes list. Available filters:

dangling=<boolean> When set to true (or 1), returns all volumes that are not in use by a container. When set to false (or 0), only volumes that are in use by one or more containers are returned.

driver=<volume-driver-name> Matches volumes based on their driver.

label=<key> or label=<key>:<value> Matches volumes based on the presence of a label alone or a label and a value.

name=<volume-name> Matches all or part of a volume name.

volume_prune

Delete unused volumes. Similar to the cli command docker volume prune. Usage:
  volume_prune(filters = NULL)
Arguments:

filters: Filters to process on the prune list, as a named character vector.

Available filters:

volume_delete

Remove a volume. Instruct the driver to remove the volume. Similar to the cli command docker volume rm. Usage:
  volume_delete(name, force = NULL)
Arguments:

name: Volume name or ID

force: Force the removal of the volume

See also

docker_volume for information on volume objects.